requirements.txt
# Web framework & server
fastapi==0.115.0
uvicorn[standard]==0.30.0

# HTTP client (used internally by LangChain/OpenAI)
httpx==0.27.0

# Environment variables
python-dotenv==1.0.1

# Data validation (pin to v2 — compatible with FastAPI 0.115)
pydantic==2.7.4

# LangChain stack
langchain==0.3.0
langchain-core==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0